+++ src/Svelte/Components/Player/Profile.svelte
@@ -4,12 +4,14 @@
     import ProfilePpCalc from './ProfilePpCalc.svelte';
     import Button from '../Common/Button.svelte';
     import Badge from "../Common/Badge.svelte";
+    import Select from "../Common/Select.svelte";
 
     import {getConfig} from '../../../plugin-config';
     import {UNRANKED} from '../../../scoresaber/rankeds';
     import {diffColors} from '../../../song';
     import eventBus from '../../../utils/broadcast-channel-pubsub';
-    import {_} from '../../stores/i18n';
+    import {_, trans} from '../../stores/i18n';
+    import {dateFromString} from "../../../utils/date";
 
     export let profile;
 
@@ -17,15 +19,44 @@
     let showCalc = false;
     let showBadges = true;
 
+    const ALL = 365 * 100;
+    let strings = {
+        periods: [
+            {_key: 'dashboard.periods.last3Days', value: 3},
+            {_key: 'dashboard.periods.lastWeek', value: 7},
+            {_key: 'dashboard.periods.last2Weeks', value: 14},
+            {_key: 'dashboard.periods.lastMonth', value: 30},
+            {_key: 'dashboard.periods.last3Months', value: 90},
+            {_key: 'dashboard.periods.lastHalfYear', value: 182},
+            {_key: 'dashboard.periods.lastYear', value: 365},
+            {_key: 'dashboard.periods.all', value: ALL},
+        ]
+    }
+
+    let values = {
+        selectedPeriod: strings.periods.find(p => p.value === ALL)
+    }
+
+    function translateAllStrings() {
+        Object.keys(strings).forEach(key => {
+            strings[key].forEach(item => {
+                if (item._key) item.label = trans(item._key);
+            })
+        })
+
+        strings = {...strings};
+        values = {...values};
+    }
+
     const badgesDef = [
-        {name: 'A', min: null, max: 80, value: 0, color: diffColors.easy},
-        {name: 'S', min: 80, max: 85, value: 0, color: diffColors.normal},
-        {name: 'S+', min: 85, max: 90, value: 0, color: diffColors.hard},
-        {name: 'SS', min: 90, max: 95, value: 0, color: diffColors.expert},
         {name: 'SS+', min: 95, max: null, value: 0, color: diffColors.expertPlus},
+        {name: 'SS', min: 90, max: 95, value: 0, color: diffColors.expert},
+        {name: 'S+', min: 85, max: 90, value: 0, color: diffColors.hard},
+        {name: 'S', min: 80, max: 85, value: 0, color: diffColors.normal},
+        {name: 'A', min: null, max: 80, value: 0, color: diffColors.easy},
     ];
 
-    onMount(async() => {
+    onMount(async () => {
         const profileConfig = await getConfig('profile');
         if (profileConfig && profileConfig.showOnePpCalc) showCalc = true;
 
@@ -43,56 +74,68 @@
     })
 
     function getPlayerStats(scores) {
-        if (!scores || !scores.length) return null;
+        badgesDef.forEach(badge => {
+            badge.value = 0;
+            badge.title = !badge.min ? '< ' + badge.max + '%' : (!badge.max ? '> ' + badge.min + '%' : badge.min + '% - ' + badge.max + '%');
+        });
 
-        const stats = scores.reduce((cum, s) => {
+        let stats = {badges: badgesDef, totalAcc: 0, totalScore: 0, avgAcc: 0, playCount: scores.length, medianAcc: 0, stdDeviation: 0};
+
+        if (!scores || !scores.length) return stats;
+
+        stats = scores.reduce((cum, s) => {
             if (!s.maxScoreEx) return cum;
 
             const scoreMult = s.uScore ? s.score / s.uScore : 1
-            const acc = s.score / s.maxScoreEx / scoreMult * 100;
+            s.acc = s.score / s.maxScoreEx / scoreMult * 100;
 
             cum.totalScore += s.uScore ? s.uScore : s.score;
-            cum.totalAcc += acc;
+            cum.totalAcc += s.acc;
 
             cum.badges.forEach(badge => {
-                if ((!badge.min || badge.min <= acc) && (!badge.max || badge.max > acc)) badge.value++;
-
-                badge.title = !badge.min ? '< ' + badge.max + '%' : (!badge.max ? '> ' + badge.min + '%' : badge.min + '% - ' + badge.max + '%')
+                if ((!badge.min || badge.min <= s.acc) && (!badge.max || badge.max > s.acc)) badge.value++;
             })
 
             return cum;
         }, {badges: badgesDef, totalAcc: 0, totalScore: 0, avgAcc: 0, playCount: scores.length})
 
+        stats.medianAcc = (scores.sort((a, b) => a.acc - b.acc))[Math.ceil(scores.length / 2)].acc;
         stats.avgAcc = stats.totalAcc / scores.length;
+        stats.stdDeviation = Math.sqrt(scores.reduce((sum, s) => sum + Math.pow(stats.avgAcc - s.acc, 2), 0) / scores.length);
+
         delete stats.totalAcc;
 
         return stats;
     }
 
     $: allRankedScores = profile.scores
-            ? Object.values(profile.scores)
-                    .filter(s => s.pp > 0 && !UNRANKED.includes(s.leaderboardId))
-            : null;
+      ? Object.values(profile.scores)
+        .filter(s => s.pp > 0 && !UNRANKED.includes(s.leaderboardId))
+      : null;
 
-    $: scores = allRankedScores
-            ? allRankedScores
-                    .map(s => s.pp)
-                    .sort((a, b) => b - a)
-            : null;
+    $: filteredScores = allRankedScores.filter(s => values.selectedPeriod.value === ALL || Date.now() - dateFromString(s.timeset).getTime() <= values.selectedPeriod.value * 1000 * 60 * 60 * 24)
 
-    $: stats = getPlayerStats(allRankedScores);
+    $: stats = getPlayerStats(filteredScores);
+
+    $: {
+        translateAllStrings($_);
+    }
 </script>
 
 {#if profile && stats}
-    <ProfileLine label={$_.profile.stats.rankedPlayCount} value={stats.playCount} precision={0}/>
+    <ProfileLine label={$_.profile.stats.rankedPlayCount} value={stats.playCount} precision={0}>
+        <Select bind:value={values.selectedPeriod} items={strings.periods} right={true}/>
+    </ProfileLine>
     <ProfileLine label={$_.profile.stats.totalRankedScore} value={stats.totalScore} precision={0}/>
     <ProfileLine label={$_.profile.stats.avgRankedAccuracy} value={stats.avgAcc} suffix="%"/>
+    <ProfileLine label={$_.profile.stats.stdDeviationRankedAccuracy} value={stats.stdDeviation} suffix="%"/>
+    <ProfileLine label={$_.profile.stats.medianRankedAccuracy} value={stats.medianAcc} suffix="%"/>
 {/if}
 
-{#if showCalc && scores && scores.length}
+{#if showCalc && allRankedScores && allRankedScores.length}
     <li class="calc">
         <div>
-            <ProfilePpCalc scores={scores} playerId={profile.id} mode={mode}/>
+            <ProfilePpCalc scores={allRankedScores} playerId={profile.id} mode={mode}/>
         </div>
         <Button iconFa="fas fa-arrows-alt-v" on:click={() => mode = mode === 'pp-stars' ? 'stars-pp' : 'pp-stars'}/>
     </li>
@@ -100,13 +143,19 @@
 
 {#if showBadges && stats && stats.badges}
     <div class="badges">
-        {#each stats.badges.reverse() as badge (badge)}
+        {#each stats.badges as badge (badge)}
             <Badge name={badge.name} value={badge.value} title={badge.title} color="white" bgColor={badge.color} digits={0}/>
         {/each}
     </div>
 {/if}
 
 <style>
+    :global(li .multi-select) {
+        position: absolute;
+        top: -.5rem;
+        right: 0;
+    }
+
     li.calc {
         display: flex;
         align-items: flex-start;